Separate base alert component from bus shelter specific parts
